odoo/requirements.txt
# requirements.txt
# Layered Python dependencies for Odoo 19 + OCA + AI
#
# Structure:
#   requirements-odoo.txt  — Odoo CE core (placeholder; baked into image)
#   requirements-oca.txt   — extra libs needed by installed OCA modules
#   requirements-ai.txt    — OCA/ai bridge + ipai_ai_* module deps
#
# Install: pip install -r requirements.txt

-r requirements-odoo.txt
-r requirements-oca.txt
-r requirements-ai.txt

# ----- Pre-commit & Quality Tools -----
pre-commit==4.5.1
black==26.1.0
isort==7.0.0
flake8==7.3.0
yamllint==1.38.0

# ----- OCA Dependency Management -----
git-aggregator==4.1

# ----- OCA Linting -----
pylint>=3.0.0,<4.1.0
pylint-odoo==10.0.1

# ----- Testing -----
pytest==9.0.2
pytest-cov==7.0.0
coverage==7.13.4

# ----- Debugging -----
debugpy==1.8.20
ipython==8.25.0

# ----- Documentation -----
sphinx==9.1.0
sphinx-rtd-theme==2.0.0

# ----- Odoo ORM Stubs (for IDE autocomplete) -----
# Uncomment if using VS Code with Pylance
# odoo-stubs==18.0.0

# ===========================================
# Cloud IDE / CI Runner Tooling
# ===========================================
# For Supabase migrations checks, edge-function
# webhook checks via HTTP, DB introspection,
# ERD export via DBML, and general ops scripting.
# NOT Odoo-runtime (Odoo runs in its own container).
# ===========================================

# ----- Core Scripting -----
python-dotenv==1.2.2
pydantic==2.9.2
pydantic-settings==2.13.1
requests==2.32.5
httpx==0.28.1
tenacity==9.1.4
rich==14.3.3
pyyaml==6.0.3
jinja2==3.1.6

# ----- Postgres / Supabase DB Ops -----
psycopg[binary]==3.3.3
sqlalchemy==2.0.46

# ----- Schema / ERD / Docs Generation -----
dbml-to-sqlalchemy==0.9.0
sqlalchemy_schemadisplay==2.0
graphviz==0.21

# ----- Schema Validation -----
jsonschema==4.26.0

# ----- CLI Ergonomics -----
typer==0.23.1
